+++ Changelog.md
@@ -1177,3 +1177,9 @@ A blank line will now shown and deletes immediately when the command are all fin
 # 20251016 (*)
 
 # Rewrited the entire Music Player system
+
+#
+
+# 20251022
+
+(-/+) Refactor all history from old repo