+++ Changelog.md
@@ -1183,3 +1183,5 @@ A blank line will now shown and deletes immediately when the command are all fin
 # 20251022
 
 (-/+) Refactor all history from old repo
+
+(+) Minior code opitmization